Unified testing utils, part 1
