+++ tests/testing_utils/files_and_dirs.py
@@ -77,29 +77,3 @@ def temp_local_dir(
                         fh.write(contents)
 
         yield Path(tmpdir)
-
-
-def merge_left(target: Dict, source: Dict) -> None:
-    """
-    Recursively merges key/value pairs from source into target.
-    Modifies the original dict-like "target".
-    """
-    for k, v in source.items():
-        if k in target and isinstance(target[k], dict):
-            # assumption: all inputs have been validated.
-            merge_left(target[k], v)
-        else:
-            target[k] = v
-
-
-# Borrowed from tests_integration/test_utils.py
-# TODO: remove from here when testing utils become shared
-# TODO: contextlib.chdir isn't available before Python 3.11, so this is an alternative for older versions
-@contextmanager
-def pushd(directory: Path):
-    cwd = os.getcwd()
-    os.chdir(directory)
-    try:
-        yield directory
-    finally:
-        os.chdir(cwd)
