+++ tests_common/path_utils.py
@@ -11,9 +11,15 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+from __future__ import annotations
 
+import contextlib
+import os
+import tempfile
 import unittest.mock as mock
-from pathlib import PurePosixPath
+from contextlib import contextmanager
+from pathlib import PurePosixPath, Path
+from tempfile import TemporaryDirectory
 
 import pytest
 
@@ -29,3 +35,66 @@ def print_paths_as_posix():
     with mock.patch("pathlib.WindowsPath.__str__", autospec=True) as mock_str:
         mock_str.side_effect = lambda path: str(PurePosixPath(*path.parts))
         yield mock_str
+
+
+@contextmanager
+def pushd(directory: Path):
+    cwd = os.getcwd()
+    os.chdir(directory)
+    try:
+        yield directory
+    finally:
+        os.chdir(cwd)
+
+
+@contextmanager
+def _named_temporary_file(suffix=None, prefix=None):
+    with tempfile.TemporaryDirectory() as tmp_dir:
+        suffix = suffix or ""
+        prefix = prefix or ""
+        f = Path(tmp_dir) / f"{prefix}tmp_file{suffix}"
+        f.touch()
+        yield f
+
+
+@pytest.fixture()
+def named_temporary_file():
+    return _named_temporary_file
+
+
+class WorkingDirectoryChanger:
+    def __init__(self):
+        self._initial_working_directory = os.getcwd()
+
+    @staticmethod
+    def change_working_directory_to(directory: str | Path):
+        os.chdir(directory)
+
+    def restore_initial_working_directory(self):
+        self.change_working_directory_to(self._initial_working_directory)
+
+
+@pytest.fixture
+def temporary_working_directory():
+    working_directory_changer = WorkingDirectoryChanger()
+    with TemporaryDirectory() as tmp_dir:
+        working_directory_changer.change_working_directory_to(tmp_dir)
+        yield Path(tmp_dir)
+        working_directory_changer.restore_initial_working_directory()
+
+
+# TODO: remove, alias for other fixture
+temp_dir = temporary_working_directory
+
+
+@pytest.fixture
+def temporary_working_directory_ctx():
+    @contextlib.contextmanager
+    def _ctx_manager():
+        working_directory_changer = WorkingDirectoryChanger()
+        with TemporaryDirectory() as tmp_dir:
+            working_directory_changer.change_working_directory_to(tmp_dir)
+            yield Path(tmp_dir)
+            working_directory_changer.restore_initial_working_directory()
+
+    return _ctx_manager
