+++ tests_integration/nativeapp/test_teardown.py
@@ -14,7 +14,7 @@
 from shlex import split
 
 from tests.project.fixtures import *
-from tests_integration.test_utils import (
+from tests_integration.testing_utils.cursor_helpers import (
     contains_row_with,
     not_contains_row_with,
     row_from_snowflake_session,
