+++ tests_integration/testing_utils/__init__.py
@@ -17,4 +17,4 @@ from tests_integration.testing_utils.assertions.test_result_assertions import *
 from tests_integration.testing_utils.naming_utils import *
 from tests_integration.testing_utils.snowpark_utils import *
 from tests_integration.testing_utils.sql_utils import *
-from tests_integration.testing_utils.working_directory_utils import *
+from tests_common.path_utils import *
